Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/webaudio/the-audio-api/the-waveshapernode-interface/waveshaper-copy-curve-expected.txt

CONSOLE MESSAGE: line 31: SyntaxError: The string did not match the expected pattern.

Harness Error (FAIL), message = SyntaxError: The string did not match the expected pattern.

PASS # AUDIT TASK RUNNER STARTED. 
PASS > [test copying] Modifying curve should not modify WaveShaper 
